requirements.txt
flask
openai==0.28
Pillow
pytesseract
markdown
Flask-Session